+++ src/app/modules/feedback/__tests__/feedback.service.spec.ts
@@ -6,19 +6,14 @@ import { omit, times } from 'lodash'
 import moment from 'moment-timezone'
 import mongoose from 'mongoose'
 
-import getAdminFeedbackModel from 'src/app/models/admin_feedback.server.model'
 import getFormFeedbackModel from 'src/app/models/form_feedback.server.model'
 
 import { FormFeedbackMetaDto } from '../../../../../shared/types'
 import { DatabaseError } from '../../core/core.errors'
-import {
-  DuplicateFeedbackSubmissionError,
-  MissingAdminFeedbackError,
-} from '../feedback.errors'
+import { DuplicateFeedbackSubmissionError } from '../feedback.errors'
 import * as FeedbackService from '../feedback.service'
 
 const FormFeedback = getFormFeedbackModel(mongoose)
-const AdminFeedbackModel = getAdminFeedbackModel(mongoose)
 
 describe('feedback.service', () => {
   beforeAll(() => dbHandler.connect())
@@ -316,242 +311,6 @@ describe('feedback.service', () => {
       expect(existSpy).toHaveBeenCalledWith({
         submissionId: MOCK_SUBMISSION_ID,
       })
-      expect(actualResult.isErr()).toEqual(true)
-      expect(actualResult._unsafeUnwrapErr()).toBeInstanceOf(DatabaseError)
-    })
-  })
-
-  describe('insertAdminFeedback', () => {
-    beforeEach(async () => {
-      await dbHandler.clearCollection(AdminFeedbackModel.collection.name)
-    })
-    afterEach(() => jest.clearAllMocks())
-
-    const MOCK_USER_ID = new ObjectId().toHexString()
-    const MOCK_RATING = 1
-    const MOCK_COMMENT = 'mock comment'
-    const MOCK_ADMIN_FEEDBACK = new AdminFeedbackModel({
-      userId: MOCK_USER_ID,
-      rating: MOCK_RATING,
-      comment: MOCK_COMMENT,
-    })
-
-    it('should return Admin Feedback document on successful insertion with comment', async () => {
-      // Arrange
-      // Mock success.
-      const insertSpy = jest
-        .spyOn(AdminFeedbackModel, 'create')
-        // @ts-ignore
-        .mockResolvedValueOnce(MOCK_ADMIN_FEEDBACK)
-
-      // Act
-      const actualResult = await FeedbackService.insertAdminFeedback({
-        userId: MOCK_USER_ID,
-        rating: MOCK_RATING,
-        comment: MOCK_COMMENT,
-      })
-
-      // Assert
-      expect(insertSpy).toHaveBeenCalledTimes(1)
-      expect(actualResult.isOk()).toEqual(true)
-      expect(actualResult._unsafeUnwrap()).toEqual(MOCK_ADMIN_FEEDBACK)
-    })
-
-    it('should return Admin Feedback document on successful insertion without comment', async () => {
-      const mock_feedback_no_comment = new AdminFeedbackModel({
-        userId: MOCK_USER_ID,
-        rating: MOCK_RATING,
-      })
-      // Arrange
-      // Mock success.
-      const insertSpy = jest
-        .spyOn(AdminFeedbackModel, 'create')
-        // @ts-ignore
-        .mockResolvedValueOnce(mock_feedback_no_comment)
-
-      // Act
-      const actualResult = await FeedbackService.insertAdminFeedback({
-        userId: MOCK_USER_ID,
-        rating: MOCK_RATING,
-      })
-
-      // Assert
-      expect(insertSpy).toHaveBeenCalledTimes(1)
-      expect(actualResult.isOk()).toEqual(true)
-      expect(actualResult._unsafeUnwrap()).toEqual(mock_feedback_no_comment)
-    })
-
-    it('should return DatabaseError when error occurs whilst inserting feedback', async () => {
-      // Arrange
-      // Mock failure
-      const insertSpy = jest
-        .spyOn(AdminFeedbackModel, 'create')
-        // @ts-ignore
-        .mockRejectedValueOnce(new Error('some error'))
-
-      // Act
-      const actualResult = await FeedbackService.insertAdminFeedback({
-        userId: MOCK_USER_ID,
-        rating: MOCK_RATING,
-      })
-
-      // Assert
-      expect(insertSpy).toHaveBeenCalledTimes(1)
-      expect(actualResult.isErr()).toEqual(true)
-      expect(actualResult._unsafeUnwrapErr()).toBeInstanceOf(DatabaseError)
-    })
-  })
-
-  describe('updateAdminFeedback', () => {
-    const MOCK_FEEDBACK_ID = new ObjectId().toHexString()
-    const MOCK_USER_ID = new ObjectId().toHexString()
-    const MOCK_RATING = 1
-    const MOCK_COMMENT = 'mock comment'
-
-    beforeEach(async () => {
-      await dbHandler.clearCollection(AdminFeedbackModel.collection.name)
-      await AdminFeedbackModel.create({
-        _id: MOCK_FEEDBACK_ID,
-        userId: MOCK_USER_ID,
-        rating: MOCK_RATING,
-        comment: MOCK_COMMENT,
-      })
-    })
-    afterEach(() => jest.clearAllMocks())
-
-    it('should update feedback successfully with both rating and comment changes', async () => {
-      const newRating = 0 as number
-      const newComment = 'new comment'
-      const expectedResult = new AdminFeedbackModel({
-        userId: MOCK_USER_ID,
-        rating: newRating,
-        comment: newComment,
-      })
-
-      // Act
-      const actualResult = await FeedbackService.updateAdminFeedback({
-        feedbackId: MOCK_FEEDBACK_ID,
-        userId: MOCK_USER_ID,
-        rating: newRating,
-        comment: newComment,
-      })
-
-      const newFeedback = await AdminFeedbackModel.findById(MOCK_FEEDBACK_ID)
-
-      // Assert
-      expect(actualResult.isOk()).toEqual(true)
-      expect(newFeedback?.comment).toEqual(expectedResult.comment)
-      expect(newFeedback?.rating).toEqual(expectedResult.rating)
-    })
-
-    it('should update feedback successfully with only rating change', async () => {
-      const newRating = 0 as number
-      const expectedResult = new AdminFeedbackModel({
-        userId: MOCK_USER_ID,
-        rating: newRating,
-        comment: MOCK_COMMENT,
-      })
-
-      // Act
-      const actualResult = await FeedbackService.updateAdminFeedback({
-        feedbackId: MOCK_FEEDBACK_ID,
-        userId: MOCK_USER_ID,
-        rating: newRating,
-      })
-
-      const newFeedback = await AdminFeedbackModel.findById(MOCK_FEEDBACK_ID)
-
-      // Assert
-      expect(actualResult.isOk()).toEqual(true)
-      expect(newFeedback?.comment).toEqual(expectedResult.comment)
-      expect(newFeedback?.rating).toEqual(expectedResult.rating)
-    })
-
-    it('should update feedback successfully with only comment changes', async () => {
-      const newComment = 'new comment'
-      const expectedResult = new AdminFeedbackModel({
-        userId: MOCK_USER_ID,
-        rating: MOCK_RATING,
-        comment: newComment,
-      })
-
-      // Act
-      const actualResult = await FeedbackService.updateAdminFeedback({
-        feedbackId: MOCK_FEEDBACK_ID,
-        userId: MOCK_USER_ID,
-        comment: newComment,
-      })
-
-      const newFeedback = await AdminFeedbackModel.findById(MOCK_FEEDBACK_ID)
-
-      // Assert
-      expect(actualResult.isOk()).toEqual(true)
-      expect(newFeedback?.comment).toEqual(expectedResult.comment)
-      expect(newFeedback?.rating).toEqual(expectedResult.rating)
-    })
-
-    it('should return MissingAdminFeedbackError if feedbackId is invalid', async () => {
-      // Act
-      const actualResult = await FeedbackService.updateAdminFeedback({
-        feedbackId: new ObjectId().toHexString(),
-        userId: MOCK_USER_ID,
-        comment: 'new comment',
-      })
-
-      // Assert
-      expect(actualResult.isErr()).toEqual(true)
-      expect(actualResult._unsafeUnwrapErr()).toBeInstanceOf(
-        MissingAdminFeedbackError,
-      )
-    })
-
-    it('should return MissingAdminFeedbackError if userId is different', async () => {
-      // Act
-      const actualResult = await FeedbackService.updateAdminFeedback({
-        feedbackId: MOCK_FEEDBACK_ID,
-        userId: new ObjectId().toHexString(),
-        comment: 'new comment',
-      })
-
-      // Assert
-      expect(actualResult.isErr()).toEqual(true)
-      expect(actualResult._unsafeUnwrapErr()).toBeInstanceOf(
-        MissingAdminFeedbackError,
-      )
-    })
-
-    it('should return DatabaseError if database error occurs in findById', async () => {
-      // Arrange
-      jest
-        .spyOn(AdminFeedbackModel, 'findById')
-        .mockRejectedValueOnce(new DatabaseError())
-
-      // Act
-      const actualResult = await FeedbackService.updateAdminFeedback({
-        feedbackId: MOCK_FEEDBACK_ID,
-        userId: MOCK_USER_ID,
-        comment: 'new comment',
-      })
-
-      // Assert
-      expect(actualResult.isErr()).toEqual(true)
-      expect(actualResult._unsafeUnwrapErr()).toBeInstanceOf(DatabaseError)
-    })
-
-    it('should return DatabaseError if database error occurs in findByIdAndUpdate', async () => {
-      // Arrange
-      jest
-        .spyOn(AdminFeedbackModel, 'findByIdAndUpdate')
-        .mockRejectedValueOnce(new DatabaseError())
-
-      // Act
-      const actualResult = await FeedbackService.updateAdminFeedback({
-        feedbackId: MOCK_FEEDBACK_ID,
-        userId: MOCK_USER_ID,
-        comment: 'new comment',
-      })
-
-      // Assert
       expect(actualResult.isErr()).toEqual(true)
       expect(actualResult._unsafeUnwrapErr()).toBeInstanceOf(DatabaseError)
     })
